fix: C2 Float→Numeric型変更 + Alembic導入（001_float_to_numeric）

+++ models.py
@@ -1,5 +1,5 @@
 from datetime import datetime
-from sqlalchemy import Column, Integer, String, Float, DateTime, Date, ForeignKey, Text, Boolean
+from sqlalchemy import Column, Integer, String, Float, Numeric, DateTime, Date, ForeignKey, Text, Boolean
 from sqlalchemy.orm import relationship, DeclarativeBase
 
 
@@ -46,7 +46,7 @@ class Product(Base):
     name = Column(String(200), nullable=False)
     category = Column(String(50), nullable=False)
     sku = Column(String(100), unique=True, nullable=True)
-    unit_price = Column(Float, nullable=False)
+    unit_price = Column(Numeric(12, 2), nullable=False)
     unit = Column(String(20), nullable=True)
     stock_alert_threshold = Column(Integer, default=10)
     alert_enabled = Column(Boolean, default=True, nullable=False)
@@ -99,7 +99,7 @@ class Quote(Base):
     customer_id = Column(Integer, ForeignKey("customers.id"))
     end_user_id = Column(Integer, ForeignKey("customers.id"), nullable=True) 
     status = Column(String(20), default="draft")
-    total_amount = Column(Float, default=0.0)
+    total_amount = Column(Numeric(12, 2), default=0.0)
     valid_until = Column(Date, nullable=True)
     notes = Column(Text, nullable=True)
     created_at = Column(DateTime, default=datetime.now)
@@ -127,9 +127,9 @@ class QuoteItem(Base):
     quote_id = Column(Integer, ForeignKey("quotes.id"))
     product_id = Column(Integer, ForeignKey("products.id"))
     quantity = Column(Integer, nullable=False)
-    unit_price = Column(Float, nullable=False)
+    unit_price = Column(Numeric(12, 2), nullable=False)
     discount_rate = Column(Float, default=1.0) 
-    subtotal = Column(Float, default=0.0)
+    subtotal = Column(Numeric(12, 2), default=0.0)
     quote = relationship("Quote", back_populates="items")
     product = relationship("Product", back_populates="quote_items")
 
@@ -167,11 +167,11 @@ class Sale(Base):
     customer_id = Column(Integer, ForeignKey("customers.id"), nullable=False)
     product_id = Column(Integer, ForeignKey("products.id"), nullable=False)
     quantity = Column(Integer, nullable=False, default=1)
-    unit_price = Column(Float, nullable=False)
-    subtotal = Column(Float, nullable=False)
+    unit_price = Column(Numeric(12, 2), nullable=False)
+    subtotal = Column(Numeric(12, 2), nullable=False)
     tax_rate = Column(Float, default=0.10)
-    tax_amount = Column(Float, nullable=False)
-    total_amount = Column(Float, nullable=False)
+    tax_amount = Column(Numeric(12, 2), nullable=False)
+    total_amount = Column(Numeric(12, 2), nullable=False)
     sale_date = Column(Date, nullable=False)
     status = Column(String(20), default="confirmed")  # confirmed / invoiced / paid
     notes = Column(Text, nullable=True)
@@ -191,9 +191,9 @@ class Invoice(Base):
     customer_id = Column(Integer, ForeignKey("customers.id"), nullable=False)
     issue_date = Column(Date, nullable=False)
     due_date = Column(Date, nullable=True)
-    subtotal = Column(Float, default=0.0)
-    tax_amount = Column(Float, default=0.0)
-    total_amount = Column(Float, default=0.0)
+    subtotal = Column(Numeric(12, 2), default=0.0)
+    tax_amount = Column(Numeric(12, 2), default=0.0)
+    total_amount = Column(Numeric(12, 2), default=0.0)
     status = Column(String(20), default="unpaid")  # unpaid / partial / paid
     notes = Column(Text, nullable=True)
     staff_name = Column(String(100), nullable=True)
@@ -208,7 +208,7 @@ class InvoiceItem(Base):
     id = Column(Integer, primary_key=True, autoincrement=True)
     invoice_id = Column(Integer, ForeignKey("invoices.id"), nullable=False)
     sale_id = Column(Integer, ForeignKey("sales.id"), nullable=False)
-    amount = Column(Float, nullable=False)
+    amount = Column(Numeric(12, 2), nullable=False)
     invoice = relationship("Invoice", back_populates="items")
     sale = relationship("Sale", back_populates="invoice_items")
 
@@ -218,7 +218,7 @@ class Payment(Base):
     id = Column(Integer, primary_key=True, autoincrement=True)
     invoice_id = Column(Integer, ForeignKey("invoices.id"), nullable=False)
     payment_date = Column(Date, nullable=False)
-    amount = Column(Float, nullable=False)
+    amount = Column(Numeric(12, 2), nullable=False)
     method = Column(String(50), nullable=True)  # 銀行振込 / 現金 等
     notes = Column(Text, nullable=True)
     staff_name = Column(String(100), nullable=True)
@@ -291,7 +291,7 @@ class Repair(Base):
     sent_to_maker_date       = Column(Date, nullable=True)
     maker_response           = Column(String(20), nullable=True)
     maker_response_date      = Column(Date, nullable=True)
-    maker_quote_amount       = Column(Float, nullable=True)
+    maker_quote_amount       = Column(Numeric(12, 2), nullable=True)
     maker_response_note      = Column(Text, nullable=True)
     quote_id                 = Column(Integer, nullable=True)
     quote_submitted_date     = Column(Date, nullable=True)
@@ -322,7 +322,7 @@ class RepairRecord(Base):
     symptom = Column(Text, nullable=False)                        # 故障症状
     cause = Column(Text, nullable=True)                           # 原因（判明後）
     repair_vendor = Column(String(200), nullable=True)            # 修理業者
-    repair_cost = Column(Float, nullable=True)                    # 修理費用
+    repair_cost = Column(Numeric(12, 2), nullable=True)                    # 修理費用
     sent_date = Column(Date, nullable=True)                       # メーカー送付日
     repaired_date = Column(Date, nullable=True)                   # 修理完了・返却日
     status = Column(String(20), nullable=False, default="reported")
